load_testing/requirements.txt
locust==2.17.0
Pillow==10.0.1
requests==2.32.3
